refactor: extract Modal component
